requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
astroid==3.2.4
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
billiard==4.2.0
cachetools==5.3.1
celery==5.3.1
certifi==2023.7.22
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==43.0.0
daphne==4.1.2
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
Django==4.2.5
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-decouple==2.1
django-import-export==3.2.0
django-redis==5.3.0
django-soft-delete==0.9.21
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
frozenlist==1.4.1
google-auth==2.22.0
hyperlink==21.0.0
idna==3.4
incremental==24.7.2
inflection==0.5.1
isort==5.13.2
kombu==5.4.0
mccabe==0.7.0
msgpack==1.0.8
multidict==6.0.5
odfpy==1.4.1
openpyxl==3.1.5
packaging==23.1
passlib==1.7.4
Pillow==10.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pyfcm==2.0.6
PyJWT==2.8.0
pylint==3.2.6
pyOpenSSL==24.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
rsa==4.9
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tablib==3.6.1
tomlkit==0.13.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
zope.interface==7.0.1
